move platformjsongraph
