+++ config/packages/prod/Centreon.yaml
@@ -7,6 +7,9 @@ jms_serializer:
             application:
                 namespace_prefix: "Centreon\\Application"
                 path: '%kernel.root_dir%/../config/packages/serializer/Application'
+            infrastructure:
+                namespace_prefix: "Centreon\\Infrastructure"
+                path: '%kernel.root_dir%/../config/packages/serializer/Infrastructure'
 parameters:
     api.header: "Api-Version"
     api.version.latest: "2.0"
